01-tutorials/02-multi-agent-systems/02-swarm-agent/requirements.txt
strands-agents
strands-agents-tools
arize-otel
arize-toolkit
strands-agents[openai]
arize-otel 
openinference-instrumentation-openai 
opentelemetry-exporter-otlp
python-dotenv